Adjustable interval and amount for streakers

+++ src/constants.h
@@ -27,7 +27,7 @@ namespace constants {
     extern const byte LED_X_VALS[];
     extern const byte LED_Y_VALS[];
 
-    extern const byte LED_GRID_INDECIES[];
+    extern const int LED_GRID_INDECIES[];
 }
 
 #endif
